+++ examples/oft-hyperliquid/contracts/extensions/MyHyperLiquidComposer_Recoverable.sol
@@ -10,16 +10,14 @@ import { RecoverableComposer } from "@layerzerolabs/hyperliquid-composer/contrac
 ///      The owner can also recover tokens from the composer on hyperevm to the recovery address.
 ///      The owner can recover native tokens (HYPE) and USDC as well.
 ///
-/// @dev This contract does NOT refund dust to the receiver because we do not expect any due to truncation of sharedDecimals.
+/// @dev This contract does NOT refund dust because we do not expect any due to truncation of sharedDecimals.
 ///      In the off-chance that you have dust you would have to implement dust refunds to the receiver in:
 ///      `_transferERC20ToHyperCore` and `_transferNativeToHyperCore`
 ///
-/// @dev Disclaimer: If your token's evm total supply exceeds your asset bridge's balance when scaled to EVM, it is possible
-///      that the composer will not be able to send the tokens to the receiver address on hypercore due to bridge consumption.
+/// @dev Disclaimer: If your token's evm total supply exceeds your asset bridge's balance when scaled to EVM,
+///      it is possible that the composer will not be able to send the tokens to the receiver on hypercore.
 ///      Tokens would instead be returned to the sender address on HyperEVM. Front-end handling is recommended.
 contract MyHyperLiquidComposer_Recoverable is HyperLiquidComposer, RecoverableComposer {
-    error InvalidRecoveryAddress();
-
     /// @notice Constructor for the HyperLiquidComposer with RecoverableComposer extension
     ///
     /// @param _oft The address of the OFT
@@ -32,7 +30,5 @@ contract MyHyperLiquidComposer_Recoverable is HyperLiquidComposer, RecoverableCo
         uint64 _hlIndexId,
         int8 _assetDecimalDiff,
         address _recoveryAddress
-    ) HyperLiquidComposer(_oft, _hlIndexId, _assetDecimalDiff) RecoverableComposer(_recoveryAddress) {
-        if (_recoveryAddress == address(0)) revert InvalidRecoveryAddress();
-    }
+    ) HyperLiquidComposer(_oft, _hlIndexId, _assetDecimalDiff) RecoverableComposer(_recoveryAddress) {}
 }