+++ examples/lzapp-migration/CHANGELOG.md
@@ -1,5 +1,35 @@
 # @layerzerolabs/lzapp-migration-example
 
+## 0.3.6
+
+### Patch Changes
+
+- 7ef51d7: bump up versions for anchor, solana, rust
+
+## 0.3.5
+
+### Patch Changes
+
+- 77ba155: update links
+
+## 0.3.4
+
+### Patch Changes
+
+- 002caa2: Add support for multiple ALTs when sending OFTs from Solana, fix extra options and compose message parsing bugs
+
+## 0.3.3
+
+### Patch Changes
+
+- e415855: rename retryPayload to retryMessage
+
+## 0.3.2
+
+### Patch Changes
+
+- 5e15a8d: Add lockfiles to example packages
+
 ## 0.3.1
 
 ### Patch Changes