+++ examples/oft-adapter-aptos-move/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @layerzerolabs/oft-adapter-aptos-move-example
 
+## 1.0.10
+
+### Patch Changes
+
+- 39ddee8: remove usage of MyOFTMock
+
+## 1.0.9
+
+### Patch Changes
+
+- c44e657: update evm networks used
+
+## 1.0.8
+
+### Patch Changes
+
+- 5e15a8d: Add lockfiles to example packages
+
 ## 1.0.7
 
 ### Patch Changes