+++ examples/oft-adapter/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @layerzerolabs/oft-adapter-example
 
+## 0.5.5
+
+### Patch Changes
+
+- 39ddee8: remove usage of MyOFTMock
+
+## 0.5.4
+
+### Patch Changes
+
+- 608915a: Fix extra option parsing to avoid overflow in `sendEvm` script
+
+## 0.5.3
+
+### Patch Changes
+
+- c44e657: update evm networks used
+
+## 0.5.2
+
+### Patch Changes
+
+- 5e15a8d: Add lockfiles to example packages
+
 ## 0.5.1
 
 ### Patch Changes