+++ examples/oft-alt/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @layerzerolabs/oft-alt-example
 
+## 0.0.6
+
+### Patch Changes
+
+- 9d3e558: use manual TUP instead of hardhat-deploy for upgradeable deploy scripts
+
+## 0.0.5
+
+### Patch Changes
+
+- c44e657: update evm networks used
+
+## 0.0.4
+
+### Patch Changes
+
+- 5e15a8d: Add lockfiles to example packages
+
 ## 0.0.3
 
 ### Patch Changes