+++ examples/oft-upgradeable/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @layerzerolabs/oft-upgradeable-example
 
+## 0.3.7
+
+### Patch Changes
+
+- 9d3e558: use manual TUP instead of hardhat-deploy for upgradeable deploy scripts
+
+## 0.3.6
+
+### Patch Changes
+
+- 608915a: Fix extra option parsing to avoid overflow in `sendEvm` script
+
+## 0.3.5
+
+### Patch Changes
+
+- c44e657: update evm networks used
+
+## 0.3.4
+
+### Patch Changes
+
+- 5e15a8d: Add lockfiles to example packages
+
 ## 0.3.3
 
 ### Patch Changes