+++ examples/oft/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @layerzerolabs/oft-example
 
+## 0.7.10
+
+### Patch Changes
+
+- 39ddee8: remove usage of MyOFTMock
+
+## 0.7.9
+
+### Patch Changes
+
+- 608915a: Fix extra option parsing to avoid overflow in `sendEvm` script
+
+## 0.7.8
+
+### Patch Changes
+
+- c44e657: update evm networks used
+
+## 0.7.7
+
+### Patch Changes
+
+- 5e15a8d: Add lockfiles to example packages
+
 ## 0.7.6
 
 ### Patch Changes