+++ examples/ovault-evm/CHANGELOG.md
@@ -1,5 +1,41 @@
 # @layerzerolabs/ovault-evm-example
 
+## 0.0.11
+
+### Patch Changes
+
+- 04768d9: update natspec tags to use custom prefix
+
+## 0.0.10
+
+### Patch Changes
+
+- 057f68d: composer sync 1.1 + native token support
+
+## 0.0.9
+
+### Patch Changes
+
+- 608915a: Fix extra option parsing to avoid overflow in `sendEvm` script
+
+## 0.0.8
+
+### Patch Changes
+
+- 286c350: Update lockfiles
+
+## 0.0.7
+
+### Patch Changes
+
+- 125ac8c: fix DebugSendScript to include composeValue in composeMsg
+
+## 0.0.6
+
+### Patch Changes
+
+- 5e15a8d: Add lockfiles to example packages
+
 ## 0.0.5
 
 ### Patch Changes