+++ packages/devtools-move/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @layerzerolabs/devtools-move
 
+## 1.0.14
+
+### Patch Changes
+
+- 84d469f: update packages for non-evm support
+
+## 1.0.13
+
+### Patch Changes
+
+- f27462a: baseX converter does not depend on chainType
+
+## 1.0.12
+
+### Patch Changes
+
+- db3a1d4: Updates simple config to support blocked messageLib on Solana and EVM
+
 ## 1.0.11
 
 ### Patch Changes