+++ packages/hyperliquid-composer/CHANGELOG.md
@@ -1,5 +1,44 @@
 # @layerzerolabs/hyperliquid-composer
 
+## 2.0.1
+
+### Patch Changes
+
+- f3783c4: patch computing fee
+
+## 2.0.0
+
+### Patch Changes
+
+- Updated dependencies [b858717]
+  - @layerzerolabs/oapp-evm@0.4.0
+
+## 1.0.5
+
+### Patch Changes
+
+- 48c9d7f: sdk improvements - add USDH, ci mode for set-block, and default --network to mainnet
+- d512816: new activation to activate users with the erc20 tokens that they are bridging across
+
+## 1.0.4
+
+### Patch Changes
+
+- 52ad590: readme hyperliquid docs update
+- 1b91fd2: Fix core spot command
+
+## 1.0.3
+
+### Patch Changes
+
+- 97f506f: update register-spot command and add in commands to track spotMarket gas and current deployment for a token
+
+## 1.0.2
+
+### Patch Changes
+
+- 90417ab: adding in freeze and quote asset during hip-1 deployment
+
 ## 1.0.1
 
 ### Patch Changes